requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
edge-tts==7.0.1
eval_type_backport==0.2.2
eventlet==0.39.1
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.6.0
fsspec==2025.3.2
gql==3.5.2
graphql-core==3.2.4
greenlet==3.2.0
gTTS==2.5.4
h11==0.14.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.1.0
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==78.1.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
srt==3.5.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
together==1.5.5
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
